Models/response_translation_text.py
from pydantic import BaseModel
from typing import List

class Translation(BaseModel):
    translated_text: str
    source_language: str
    target_language: str

class TranslationResponse(BaseModel):
    translations: List[Translation]
