requirements.txt
python-dotenv==0.21.1
loguru==0.6.0
uuid
fastapi[standard]==0.115.8
google-cloud-translate==2.0.1
uvicorn